// type/Reader.type.ts
export type Reader = {
  readerId: string;
  firstName: string;
  lastName: string;
  image: string;
  phone: string;
  gender: string;
  introduction: string;
  experienceYear: string;
  dateOfBirth: string;
  memberShip: string;
  expertise: string;
  quote: string;
  experience: string;
  rating: string;
  linkUrl: string;
};